Fix database models:  * Update Property Price and Area to int  * Update PropertyType to have SortRank property  * Update PropertyAggregate to have nullable AveragePrice and AveragePricePerSqM  * Add DataConstants for Property form  * Add migration

+++ PropertyAds/PropertyAds.WebApp/Data/Models/Property.cs
@@ -11,9 +11,9 @@
         [Required]
         public string Id { get; set; } = Guid.NewGuid().ToString();
 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
-        public decimal Area { get; set; }
+        public int Area { get; set; }
 
         public int Floor { get; set; }
